+++ packages/plans/package.json
@@ -2,7 +2,7 @@
   "name": "@openstatus/plans",
   "version": "1.0.0",
   "description": "",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {},
   "dependencies": {
     "@openstatus/db": "workspace:*",
