+++ packages/plans/src/index.ts
@@ -1,0 +1,3 @@
+export { allPlans } from "./config";
+export { pricingTableConfig } from "./pricing-table";
+export { getLimit } from "./utils";
