+++ packages/plans/src/types.ts
@@ -1,0 +1,21 @@
+import type { MonitorPeriodicity } from "@openstatus/db/src/schema";
+
+export type Limits = {
+  // monitors
+  monitors: number;
+  periodicity: Partial<MonitorPeriodicity>[];
+  "multi-region": boolean;
+  "data-retention": string;
+  // status pages
+  "status-pages": number;
+  "status-subscribers": boolean;
+  "custom-domain": boolean;
+  "white-label": boolean;
+  // alerts
+  notifications: boolean;
+  sms: boolean;
+  "notification-channels": number;
+  // collaboration
+  members: "Unlimited" | number;
+  "audit-log": boolean;
+};
